implemented font.Length and changed to return float

+++ engine.go
@@ -53,7 +53,7 @@ type audio interface {
 }
 
 type Font interface {
-	Length(string) uint
+	Length(string) float32
 	Render(x1, y1, x2, y2 float64, text string)
 }
